Move more code for kind/deprecation
